third_party/WebKit/LayoutTests/fast/dom/shadow/remove-shadowroot-from-document-and-destroy-crash-expected.txt

CONSOLE ERROR: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'sha1-eYyYGmKWdhpUewohaXk9o8IaLSw='". Either the 'unsafe-inline' keyword, a hash ('sha256-47DEQpj8HBSa-_TImW-5JCeuQeRkm5NMpJWZG3hSuFU='), or a nonce ('nonce-...') is required to enable inline execution.

PASS
